Move scripts into preloads; remove insecure flags
